create methods in cycles reducers
