+++ 02-ignite-timer/src/reducers/cycles/actions.ts
@@ -3,7 +3,7 @@ import { Cycle } from "./reducer";
 export enum ActionTypes {
   ADD_NEW_CYCLE = "ADD_NEW_CYCLE",
   INTERRUPT_CURRENT_CYCLE = "INTERRUPT_CURRENT_CYCLE",
-  MARK_CURRENT_CYCLES_AS_FINISHED = "MARK_CURRENT_CYCLES_AS_FINISHED",
+  MARK_CURRENT_CYCLE_AS_FINISHED = "MARK_CURRENT_CYCLE_AS_FINISHED",
 }
 
 export function addNewCycleAction(newCycle: Cycle) {
@@ -23,6 +23,6 @@ export function interruptCurrentCycleAction() {
 
 export function markCurrentCycleAsFinishedAction() {
   return {
-    type: ActionTypes.MARK_CURRENT_CYCLES_AS_FINISHED,
+    type: ActionTypes.MARK_CURRENT_CYCLE_AS_FINISHED,
   };
 }
