+++ 02-ignite-timer/src/reducers/cycles/reducer.ts
@@ -1,6 +1,11 @@
 import { produce } from "immer";
 
 import { ActionTypes } from "./actions";
+import {
+  addNewCycleMethod,
+  interruptCurrentCycleMethod,
+  markCurrentCycleAsFinishedMethod,
+} from "./methods";
 
 export type Cycle = {
   id: string;
@@ -11,46 +16,28 @@ export type Cycle = {
   finishedDate?: Date;
 };
 
-interface ICyclesState {
+export interface ICyclesState {
   cycles: Cycle[];
   activeCycleId: string | null;
 }
 
-export function cyclesReducer(state: ICyclesState, action: any) {
+export interface IActionType {
+  type: ActionTypes;
+  payload: {
+    newCycle: Cycle;
+  };
+}
+
+export function cyclesReducer(state: ICyclesState, action: IActionType) {
   switch (action.type) {
     case ActionTypes.ADD_NEW_CYCLE:
-      return produce(state, (draft) => {
-        draft.cycles.push(action.payload.newCycle);
-        draft.activeCycleId = action.payload.newCycle.id;
-      });
-    case ActionTypes.INTERRUPT_CURRENT_CYCLE: {
-      const currentCycleIndex = state.cycles.findIndex(
-        (cycle) => cycle.id === state.activeCycleId
-      );
+      return addNewCycleMethod(state, action);
 
-      if (currentCycleIndex < 0) {
-        return state;
-      }
+    case ActionTypes.INTERRUPT_CURRENT_CYCLE:
+      return interruptCurrentCycleMethod(state);
 
-      return produce(state, (draft) => {
-        draft.activeCycleId = null;
-        draft.cycles[currentCycleIndex].interruptedDate = new Date();
-      });
-    }
-    case ActionTypes.MARK_CURRENT_CYCLES_AS_FINISHED: {
-      const currentCycleIndex = state.cycles.findIndex(
-        (cycle) => cycle.id === state.activeCycleId
-      );
-
-      if (currentCycleIndex < 0) {
-        return state;
-      }
-
-      return produce(state, (draft) => {
-        draft.activeCycleId = null;
-        draft.cycles[currentCycleIndex].finishedDate = new Date();
-      });
-    }
+    case ActionTypes.MARK_CURRENT_CYCLE_AS_FINISHED:
+      return markCurrentCycleAsFinishedMethod(state);
 
     default:
       return state;
